introduce formattedleaderboard class
